Allow unvoting an item (server)

+++ server/functions/graphql/schema/resolvers/mutations/unvote.js
@@ -4,16 +4,19 @@ module.exports = async function vote(_, args, context) {
   const {pollId, itemId, voteValue} = args
   
   try {
-    const voterId = await context.getAuthUserId()
-    
-    await db.transaction(`edges/poll_item/${pollId}/${itemId}`, oldScore => oldScore - voteValue)
-    await db.remove(`edges/voter_poll_item/${voterId}/${pollId}`)
-    await db.transaction(`polls/${pollId}/totalVotes`, count => count - 1)
-    
-    const currentScore = await db.read(`edges/poll_item/${pollId}/${itemId}`)
-    await db.write(`items/${itemId}/score`, currentScore)
+    const unvoterId = await context.getAuthUserId()
+    const voterId = await db.read(`edges/voter_poll_item/${voterId}/${pollId}`)
 
-    // TODO: if voter list is visible, delete that ID
+    if (voterId === unvoterId) {
+      await db.transaction(`edges/poll_item/${pollId}/${itemId}`, oldScore => oldScore - voteValue)
+      await db.remove(`edges/voter_poll_item/${voterId}/${pollId}`)
+      await db.transaction(`polls/${pollId}/totalVotes`, count => count - 1)
+      
+      const currentScore = await db.read(`edges/poll_item/${pollId}/${itemId}`)
+      await db.write(`items/${itemId}/score`, currentScore)
+
+      // TODO: if voter list is visible, delete that ID
+    }
   }
   catch(err) {
     return Promise.reject(err)
